refactor: move PoolKey to type
