+++ packages/v4-sdk/src/utils/encodePoolParameters.ts
@@ -1,30 +1,12 @@
 import { concat, encodePacked, pad } from 'viem'
-import { Bytes32, HooksRegistration } from '../types'
+import type { BinPoolParameter, Bytes32, CLPoolParameter } from '../types'
 import { encodeHooksRegistration } from './encodeHooksRegistration'
-
-export type CLPoolParameter = {
-  /**
-   * Hooks registration for the pool
-   * @see {@link HooksRegistration}
-   */
-  hooksRegistration?: HooksRegistration
-  tickSpacing: number
-}
 
 export const encodeCLPoolParameters = (params: CLPoolParameter): Bytes32 => {
   const hooks = encodeHooksRegistration(params?.hooksRegistration)
   const tickSpacing = encodePacked(['int24'], [params.tickSpacing])
 
   return pad(concat([tickSpacing, hooks]))
-}
-
-export type BinPoolParameter = {
-  /**
-   * Hooks registration for the pool
-   * @see {@link HooksRegistration}
-   */
-  hooksRegistration?: HooksRegistration
-  binStep: number
 }
 
 export const encodeBinPoolParameters = (params: BinPoolParameter): Bytes32 => {
