+++ packages/v4-sdk/src/utils/getPoolId.ts
@@ -1,8 +1,14 @@
 import { encodeAbiParameters, keccak256, parseAbiParameters, zeroAddress } from 'viem'
-import { Bytes32 } from '../types'
+import { Bytes32, PoolKey } from '../types'
 import { encodePoolParameters } from './encodePoolParameters'
-import { PoolKey } from './poolKey'
 
+/**
+ * `PoolId` is a bytes32 of `keccak256(abi.encode(poolKey))`
+ * @see {@link PoolKey}
+ * @see {@link https://github.com/pancakeswap/pancake-v4-core/blob/main/src/types/PoolId.sol|v4-core}
+ * @param param0
+ * @returns
+ */
 export const getPoolId = ({
   currency0,
   currency1,
