+++ packages/v4-sdk/src/utils/index.ts
@@ -1,8 +1,8 @@
 export * from './convertBinIdsToRelative'
 export * from './encodeHooksRegistration'
+export * from './encodePoolKey'
 export * from './encodePoolParameters'
 export * from './getAddLiquidityParameters'
 export * from './getBinIds'
 export * from './getPoolId'
 export * from './math'
-export * from './poolKey'
